[fix] backup.sh, backup logic
the implementation for a single 'master' backup set to link all
interval types' sets against was flawed as the master set itself
was being erroneously synced to the sources as an initial step.
given all common files for all sets were tied to this set, this
meant that any changes to files were propagated throughout all
sets. this bug didn't affect new and deleted files so for those
the sets appeared correct and hence how this almighty bug went
unnoticed for so long

the fix is to ensure the master set is moved aside initially,
'master.tmp', maintaining its links to the entire backup set, but
then, an entirely new representation of the current sources state
needs to be pulled in with a 'link-dest' of this 'master.tmp'
folder. the temp folder can then be totally discarded. this way,
as intended, all common (identical) files remain tied to the
master copy, but crucially, any changed files will be stand alone
in master, with their histories still hardlinked across all backup
sets