[fix] git.sh, fnCommit 'false positive'
-git log's <path> arg appears to recognise a partial string match
of the HEAD commit's description. feels like a bug in the binary.
this resulted in only a single match being found for a string, which
in turn put git.sh's log command in a different mode than expected
(e.g. it listed all commits back from the single match as opposed to
the (expected) multiple matches only)