[fix] func_common.sh, workaround vared issue
-inconsistent editing of the buffer
-workaround forces the zle call (vared) through an interactive zsh
-without this in place it's as if the chars in the edit buffer
copid from the target var are 'locked' and stay that way until some
random combination of input chars and line navigation with sporadic
backspace and delete key thrown in for good measure causes the
characters to be editable
-'stty -a' just prior to point of entry and it appears identical to a
working interactive shell session. a minimal example to demonstrate
the behaviour is:

  zsh -c 'x=10; vared x'  # 'uneditable'
  zsh -i -c 'x=10; vared x'  # editable

given no customisations to zle widgets / env i.e. a vanilla zshrc
template, the only avenue left would be debugging the zsh binary.
i've no time for that atm