[mod] nixTools, refactor escapes
-move to '\033' and explicit transformation to escape code to avoid
various issues with displaying the code via terminal (e.g. displaying
modified code containing escapes via git diff, interprets the escapes
as it needs to for colour highlighting. this was regularly switching
the cursor off, as the 'hide' sequence happened to be below the 'show'
sequence. avoid all such shenanigans, even if we have to work harder to
force the sequences interpretation now
-upper case names
-meaningful prefixes, e.g. TERM, LN, CUR, CLR
-comments for necessary decriptions