[mod] file.sh, rename, custom replacements
-use 'custom string' replacement by default for X=Y transforms as
opposed to the original 'custom char' replacement which assumed the
source / from element of the pair (e.g. X) was a character set, and
implicitly wrapped in '[]' brackets. this seems less intuitive than
X=Y string replacement so is replaced as the default
-to recover the original behaviour, the square brackets must be
explicitly given as part of the source element, e.g. '[abc]=', which
replaces any instance of the characters 'a', 'b' or 'c' with nothing
-help updated to reflect this breaking change