[fix] backup.sh, stripping of inline comments
-switch from sed to shell expansions. the only way to distinguish
inline comments from legitimate uses of hashes in strings (e.g. "c#")
is to require a 'special' prefix, with whitespace being the obvious
(and syntactically valid choice). this shifts the extraction
criteria to 'match everything up until a word (' #', an ordered set
of characters!)' which is impossible for sed as it does not support
non-greedy matching