[fix] file.sh, transform-checks bug
-premature failure and exit triggered for all use when not under
debug mode (e.g. DEBUG 0). the ultimate (desired!) failure of the
'-n' test option for the args loop results in a non-zero return
value. as the last executed statement of the function, this was
then propagated back up through the call stack killing the expected
code path prematurely. under debug level > 0, the execution of the
conditional debug statement mitigated and thus hid that behaviour
-add an explicit zero return to indicate success